requirements.txt
requests
colorama
Flask
python-dotenv
pillow